Refactor: ERD 변경에 따른 엔티티 수정, 엔티티 수정에 따른 코드 수정

+++ src/main/java/com/cruise/parkinglotto/web/dto/applicantDTO/ApplicantResponseDTO.java
@@ -11,8 +11,8 @@ public class ApplicantResponseDTO {
         private Double weightedTotalScore;
         private WinningStatus winningStatus;
         private String parkingSpaceName;
-        private Long reserveNum;
-        private Long userSeedIndex;
+        private Integer reserveNum;
+        private Integer userSeedIndex;
         private String userSeed;
         private String firstChoice;
         private String secondChoice;
